:triangular_flag_on_post::broom: Customize error pages, clean project

+++ src/Controller/LikeController.php
@@ -4,10 +4,13 @@ namespace App\Controller;
 
 use App\Entity\Like;
 use App\Entity\User;
+use App\Entity\Demand;
 use App\Form\LikeType;
 use Doctrine\ORM\EntityManager;
 use App\Repository\LikeRepository;
 use App\Repository\UserRepository;
+use App\Repository\DemandRepository;
+use App\Repository\DemandRelationRepository;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Security\Core\Security;
 use Symfony\Component\HttpFoundation\Response;
@@ -18,50 +21,66 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 class LikeController extends AbstractController
 {
     #[Route('/', name: 'app_like_index', methods: ['GET'])]
-    public function index(Request $request , LikeRepository $likeRepository, UserRepository $userRepository, Security $security): Response
+    public function index(Request $request , LikeRepository $likeRepository, UserRepository $userRepository, DemandRepository $demandRepository, DemandRelationRepository $demandRelationRepository, Security $security): Response
     {
+        
         $userLikedId = $request->query->get('id');
         $userLiked  = $userRepository->find(['id' => $userLikedId]);
 
         $userLiker = $security->getUser();
 
-        // $isLiked=false;
-
-        $likes = $likeRepository->findBy(['user_liked' => $userLiked]);
-
         $isLiked = $userLiked->isLikedByUser($userLiker);
         
-    
 
-        // foreach ($likes as $like) {
-        //     if ($userLiker == $like->getUserLiker()) {
-        //         $isLiked=true;
-        //         $theLike = $like;
-        //     }
-        // }
-  
-        if($isLiked){
-            $theLike = $likeRepository->findOneBy(['user_liked' => $userLiked, 'user_liker' => $userLiker]);
-            $likeRepository->remove($theLike, true);
+        // find the Relations of the userLiked instance if there is one 
+        $userLikedDemands = $demandRepository->findBy(['user' => $userLiked,]);
 
-        }
-        elseif(!$isLiked) {
-            $like = new Like();
-            $like->setUserLiked($userLiked);
-            $like->setUserLiker($userLiker);
-            $likeRepository->save($like, true);
-        }
-
-            
-
-        $countLiked = $likeRepository->count(['user_liked' => $userLiked]);
-
-            return $this->json([
-                'message' => 'Like supprimé',
-                'countLiked' =>  $countLiked,
-                'isLiked' =>  $isLiked,
-                ],200);
+        // find the Relations of the userLiker instance if there is one 
+        $userLikerRelations = $demandRelationRepository->findBy(['user' => $userLiker,]);
         
+
+        // Check if the connected user has already responded to the demand 
+        $related = false;
+        foreach($userLikerRelations as $userLikerRelation)
+        {
+            foreach($userLikedDemands as $userLikedDemand)
+            {
+                //If related user = user connected
+                if($userLikerRelation->getDemand() == $userLikedDemand){
+                    $related = true;
+                    break;
+                }
+            }
+        }
+        if($related)
+        {
+            if($isLiked){
+                $theLike = $likeRepository->findOneBy(['user_liked' => $userLiked, 'user_liker' => $userLiker]);
+                $likeRepository->remove($theLike, true);
+                $isLiked = false;
+
+            }
+            elseif(!$isLiked) {
+                $like = new Like();
+                $like->setUserLiked($userLiked);
+                $like->setUserLiker($userLiker);
+                $likeRepository->save($like, true);
+                $isLiked = true;
+            }
+
+                
+
+            $countLiked = $likeRepository->count(['user_liked' => $userLiked]);
+
+                return $this->json([
+                    'message' => 'Like modifié',
+                    'countLiked' =>  $countLiked,
+                    'isLiked' =>  $isLiked,
+                    ],200);
+        }
+        else{
+                throw $this->createNotFoundException('Erreur, vous devez entrer en relation avant de pouvoir liker');
+        }
     }
